app: plic: add plic shell example
This application utilizes the PLIC SHELL that has recently been added to
zephyr. This app is totally driven by the prj.conf and
mpfs_icicle.overlay files.

the prj.conf file specifies the boot hart (in this case, u54_4), and to
use the plic shell.

The overlay specifies to use an address space in ddr (outside of Linux
address space) and also disables hart 0.

Signed-off-by: Conor Paxton <conor.paxton@microchip.com>